refactor: useTimeState 훅 분리
